+++ frontend/src/app/main/refs.cljs
@@ -284,6 +284,9 @@
               (mapv (d/getf objects) shapes)))]
     (l/derived selector selected-data =)))
 
+(def workspace-focus-selected
+  (l/derived :workspace-focus-selected st/state))
+
 ;; ---- Viewer refs
 
 (def viewer-file